Rename ColorPickerDelegate to ColorDelegate
